fix(plat/xilinx/versal): resolve misra R10.3
MISRA Violation: MISRA-C:2012 R.10.3
- The value of an expression shall not be assigned to an object with a
  narrower essential type or of a different essential type category

+++ plat/xilinx/versal/plat_versal.c
@@ -17,5 +17,5 @@ int plat_core_pos_by_mpidr(u_register_t mpidr)
 		return -1;
 	}
 
-	return versal_calc_core_pos(mpidr);
+	return (int)versal_calc_core_pos(mpidr);
 }
